fixed finger table

+++ src/csx55/chord/utils/Entry.java
@@ -2,14 +2,14 @@ package csx55.chord.utils;
 
 public class Entry {
     /* this is the ring position */
-    int start;
+    long start;
 
     /* ipAddress:port string */
     String ipAddress;
 
     int port;
 
-    public Entry(int start, String ipAddress, int port) {
+    public Entry(long start, String ipAddress, int port) {
         this.start = start;
         this.ipAddress = ipAddress;
         this.port = port;
@@ -36,7 +36,7 @@ public class Entry {
         this.port = port;
     }
 
-    public int getRingPosition() {
+    public long getRingPosition() {
         return start;
     }
